+++ src/CartMule/CartMule/AppShell.xaml.cs
@@ -1,10 +1,9 @@
-﻿namespace CartMule
+﻿namespace CartMule;
+
+public partial class AppShell : Shell
 {
-    public partial class AppShell : Shell
+    public AppShell()
     {
-        public AppShell()
-        {
-            InitializeComponent();
-        }
+        InitializeComponent();
     }
 }